# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_domain_flow.FlowLog:
    properties:
      application:
        type: string
      bytes_received:
        type: integer
      bytes_sent:
        type: integer
      dst_ip:
        type: string
      dst_port:
        type: integer
      duration:
        type: number
      id:
        type: integer
      node_id:
        type: integer
      packets_received:
        type: integer
      packets_sent:
        type: integer
      protocol:
        type: string
      recorded_at:
        type: string
      src_ip:
        type: string
      src_port:
        type: integer
    type: object
  internal_domain_flow.SavedQuery:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      query:
        type: string
    type: object
  internal_domain_ha.Config:
    properties:
      created_at:
        type: string
      heartbeat_sec:
        type: integer
      id:
        type: integer
      interface:
        type: string
      is_enabled:
        type: boolean
      mode:
        type: string
      peer_address:
        type: string
      priority:
        type: integer
      updated_at:
        type: string
      virtual_ip:
        type: string
    type: object
  internal_domain_ha.UpsertConfigRequest:
    properties:
      heartbeat_sec:
        type: integer
      interface:
        type: string
      is_enabled:
        type: boolean
      mode:
        type: string
      peer_address:
        type: string
      priority:
        type: integer
      virtual_ip:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API 支持
  description: WAF 防火墙管理系统控制面 API 接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WAF Control API
  version: "1.0"
paths:
...
  /flow-logs:
    get:
      consumes:
      - application/json
      description: 根据筛选条件分页查询流量日志，支持按源IP、目标IP、协议、时间范围过滤
      parameters:
      - description: 源IP地址
        in: query
        name: src_ip
        type: string
      - description: 目标IP地址
        in: query
        name: dst_ip
        type: string
      - description: 协议类型
        in: query
        name: protocol
        type: string
      - description: 开始时间
        in: query
        name: start_at
        type: string
      - description: 结束时间
        in: query
        name: end_at
        type: string
      - description: 排序字段
        in: query
        name: order_by
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 流量日志列表及总数
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 查询流量日志列表
      tags:
      - 流量监控
  /flow-logs/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取流量日志详情
      parameters:
      - description: 流量日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 流量日志详情
          schema:
            $ref: '#/definitions/internal_domain_flow.FlowLog'
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 未找到
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取单条流量日志
      tags:
      - 流量监控
  /flow-logs/protocols:
    get:
      consumes:
      - application/json
      description: 查询流量日志中各协议的分布情况
      produces:
      - application/json
      responses:
        "200":
          description: 协议分布数据
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取协议分布统计
      tags:
      - 流量监控
  /flow-logs/top-dst-ips:
    get:
      consumes:
      - application/json
      description: 查询流量日志中目标IP地址的Top排行
      produces:
      - application/json
      responses:
        "200":
          description: 目标IP排行数据
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取流量目标IP排行
      tags:
      - 流量监控
  /flow-logs/top-src-ips:
    get:
      consumes:
      - application/json
      description: 查询流量日志中源IP地址的Top排行
      produces:
      - application/json
      responses:
        "200":
          description: 源IP排行数据
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取流量源IP排行
      tags:
      - 流量监控
  /flow-monitor/records:
    get:
      consumes:
      - application/json
      description: 查询流量监控的汇总记录数据
      produces:
      - application/json
      responses:
        "200":
          description: 流量监控记录列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取流量监控记录
      tags:
      - 流量监控
  /flow-saved-queries:
    get:
      consumes:
      - application/json
      description: 查询所有已保存的流量查询条件
      produces:
      - application/json
      responses:
        "200":
          description: 保存的查询列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取保存的查询列表
      tags:
      - 流量监控
    post:
      consumes:
      - application/json
      description: 保存一个流量查询条件，便于后续快速复用
      parameters:
      - description: 查询条件（name和query必填）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_flow.SavedQuery'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            $ref: '#/definitions/internal_domain_flow.SavedQuery'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建保存的查询
      tags:
      - 流量监控
  /flow-saved-queries/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除已保存的流量查询条件
      parameters:
      - description: 保存的查询ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 删除成功
        "400":
          description: 无效的ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 未找到
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除保存的查询
      tags:
      - 流量监控
  /ha/config:
    get:
      consumes:
      - application/json
      description: 查询当前高可用集群的配置信息
      produces:
      - application/json
      responses:
        "200":
          description: 高可用配置详情
          schema:
            $ref: '#/definitions/internal_domain_ha.Config'
        "404":
          description: 配置未找到
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取高可用配置
      tags:
      - 高可用
    put:
      consumes:
      - application/json
      description: 创建或更新高可用集群的配置，包括模式、虚拟IP、优先级等参数
      parameters:
      - description: 高可用配置参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_domain_ha.UpsertConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 配置更新成功
          schema:
            $ref: '#/definitions/internal_domain_ha.Config'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建或更新高可用配置
      tags:
      - 高可用
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"